pyproject.toml
[tool.poetry]
name = "cyclonedds.io"
version = "0.1.0"
description = ""
authors = ["Eclipse Cyclone DDS Committers"]

[tool.poetry.dependencies]
python = "^3.8"
staticjinja = "^4.1.2"
Sphinx = "^4.4.0"
breathe = "^4.33.1"
cmake = "^3.22.2"
piccolo-theme = "^0.5.0"
ghp-import = "^2.0.2"
GitPython = "^3.1.27"
sphinx-rtd-theme = "^1.0.0"
exhale = "^0.3.1"
typing-extensions = "^4.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
